backend/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
chex==0.1.84
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
etils[epath,epy]==1.5.2
exceptiongroup==1.1.3
fastapi[standard]==0.104.1
flatbuffers==23.5.26
flax==0.7.5
fonttools==4.44.0
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
idna==3.4
importlib-resources==6.1.1
jax==0.4.20
jaxlib==0.4.20
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
markdown==3.5.1
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
nest-asyncio==1.5.8
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.2
packaging==23.2
pandas==2.1.2
pillow==10.1.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic-core==2.10.1
pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-decision-forests==1.6.0
tensorflow-estimator==2.14.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflowjs==4.12.0
tensorstore==0.1.45
termcolor==2.3.0
threadpoolctl==3.2.0
toolz==0.12.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn[standard]==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
werkzeug==3.0.1
wheel==0.41.3
wrapt==1.14.1
wurlitzer==3.0.3
zipp==3.17.0
